mainsection ing

+++ newper_shop_admin/src/main/java/com/newper/repository/MainSectionRepo.java
@@ -3,5 +3,5 @@ package com.newper.repository;
 import com.newper.entity.MainSection;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MainSectionRepo extends JpaRepository<MainSection,Integer> {
+public interface MainSectionRepo extends JpaRepository<MainSection,Long> {
 }
